Set PTS field when creating PES packets
Before the PTS was always being set to zero, which is ia violation
of the spec.  Set the PTS when generating the PES packet such that
we get PTS values within 2ms of the video frame.

Note this change depends on an updated version of libklvanc, due
to the addition of the pts argument to the library function.