Pass proper length into the vanc processing function
The original code would pass the frame stride, which isn't appropriate
given the VANC area only covers the active video (whereas the stride
covers both HANC and active video).  Also, the called function
expects the length to be specified in pixels, rather than bytes.

This problem was manifesting as VANC lines being detected three
times since the parser was overflowing into subsequent VANC lines
(in effect, the earlier lines were detecting the VANC in the subsequent
line, in addition to the actual line the VANC appeared on).