Update to support new 104->35 API in libklvanc
Move more of the business logic for 104-34 conversion into the
libklvanc library itself.  This allows us to centralize all the
logic for parsing Multiple Operation Messages and makes the
application consuming the library simpler.

Also, adjust for the reality that a Multiple Operation Message can
result in the generation of more than a single SCTE-35 section. Now
we return a list of sections which then can be fed into the mux.