When receiving SDI data in the PSF video format, treat as interlaced
Decklink Duo2 cards properly detect 1080psf/30 video, but the
drivers still expect you to refer to VANC lines using their
interlaced line numbers (i.e. refer to lines in the second field
 with numbers 570...).  In order to ensure OBE properly parses
VANC in these lines, treat the PSF video format as interlaced.

This doesn't appear to occur with the original Decklink Duo cards
because they detect streams in the PSF video format as 1080i/59,
rather than 1080psf/30.  Hence it works because it's treating it
as interlaced because it doesn't know any better.